Patch 7.5 and API 15 update

+++ NoSillyReborn/Data/PlayerData.cs
@@ -113,7 +113,7 @@ internal static class PlayerData
     #region Territory Info Tracking
 
     public static Data.TerritoryInfo? Territory { get; set; }
-    public static ushort TerritoryID => Svc.ClientState.TerritoryType;
+    public static ushort TerritoryID => (ushort)Svc.ClientState.TerritoryType;
 
     public static bool IsPvP => Territory?.IsPvP ?? false;
